==> step1_Load_LIDC.ipynb <==
import settings
import helpers
import SimpleITK  # conda install -c https://conda.anaconda.org/simpleitk SimpleITK
import numpy
import pandas
import ntpath

import cv2  # conda install -c https://conda.anaconda.org/menpo opencv3
import shutil
import random
import math
import multiprocessing
from bs4 import BeautifulSoup #  conda install beautifulsoup4, coda install lxml
import os
import glob

LUNA16_RAW_SRC_DIR ="G:/LungCancerPredict/original/luna_raw/"

LIDC_Annotations = "D:/jupyter-notebook/LungCancerPredict/LIDC_Annotations_3/"

if __name__ == "__main__":
    
     if True:
            
        process_lidc_annotations(agreement_threshold=)

def process_lidc_annotations(agreement_threshold=0):
    
    file_no = 0
    pos_count = 0
    neg_count = 0
    all_lines = []
    used_files = []
    
    for anno_dir in [d for d in glob.glob("resources/luna16_annotations/*") if os.path.isdir(d)]:

        xml_paths = glob.glob(anno_dir + "/*.xml")
        
        for xml_path in xml_paths:
            
            print(file_no, ": ",  xml_path)
            
            pos, neg, extended = load_lidc_xml(xml_path=xml_path, agreement_threshold=agreement_threshold)
            
            if pos is not None:
                
                used_files.append(xml_path)
                pos_count += len(pos)
                neg_count += len(neg)
                print("Pos: ", pos_count, " Neg: ", neg_count)
                file_no += 1
                all_lines += extended

    df_annos = pandas.DataFrame(all_lines, columns=["patient_id", "anno_index", "coord_x", "coord_y", "coord_z", "diameter", "malscore", "sphericiy", "margin", "spiculation", "texture", "calcification", "internal_structure", "lobulation", "subtlety"])
    df_annos.to_csv(LIDC_Annotations + "lidc_annotations.csv", index=False)
    

def load_lidc_xml(xml_path, agreement_threshold=0, save_nodules=False):
    
    pos_lines = []
    neg_lines = []
    extended_lines = []
    
    with open(xml_path, 'r') as xml_file:
        markup = xml_file.read()
    
    xml = BeautifulSoup(markup, features="xml")
    
    if xml.LidcReadMessage is None:
        return None, None, None
    
    patient_id = xml.LidcReadMessage.ResponseHeader.SeriesInstanceUid.text

    src_path = find_mhd_file(patient_id)
    
    if src_path is None:
        print("没有该病人CT图像")
        return None, None, None
    
    itk_img = SimpleITK.ReadImage(src_path)
    img_array = SimpleITK.GetArrayFromImage(itk_img)
    num_z, height, width = img_array.shape        #heightXwidth constitute the transverse plane
    origin = numpy.array(itk_img.GetOrigin())      # x,y,z  Origin in world coordinates (mm)
    spacing = numpy.array(itk_img.GetSpacing())    # spacing of voxels in world coor. (mm)
    rescale = spacing / settings.TARGET_VOXEL_MM
    
    reading_sessions = xml.LidcReadMessage.find_all("readingSession")
    
    #print(len(reading_sessions)) 放射学家的数量

    for reading_session in reading_sessions:

        nodules = reading_session.find_all("unblindedReadNodule")
        
        for nodule in nodules:
            
            nodule_id = nodule.noduleID.text
            rois = nodule.find_all("roi")
            
            x_min = y_min = z_min = 999999
            x_max = y_max = z_max = -999999
            
            if len(rois) < 2:
                continue
        
            for roi in rois:
                
                z_pos = float(roi.imageZposition.text)
                z_min = min(z_min, z_pos)
                z_max = max(z_max, z_pos)
                
                edge_maps = roi.find_all("edgeMap")
   
                for edge_map in edge_maps:
        
                    x = int(edge_map.xCoord.text)
                    y = int(edge_map.yCoord.text)
                    x_min = min(x_min, x)
                    y_min = min(y_min, y)
                    x_max = max(x_max, x)
                    y_max = max(y_max, y)
                    
                if x_max == x_min:
                    continue
                if y_max == y_min:
                    continue

            x_diameter = x_max - x_min
            x_center = x_min + x_diameter / 2
            y_diameter = y_max - y_min
            y_center = y_min + y_diameter / 2
            z_diameter = z_max - z_min
            z_center = z_min + z_diameter / 2
            z_center -= origin[2]
            z_center /= spacing[2]

            x_center_perc = round(x_center / img_array.shape[2], 4)
            y_center_perc = round(y_center / img_array.shape[1], 4)
            z_center_perc = round(z_center / img_array.shape[0], 4)
            
            diameter = max(x_diameter , y_diameter)
            diameter_perc = round(max(x_diameter / img_array.shape[2], y_diameter / img_array.shape[1]), 4)

            if nodule.characteristics is None:
                print("!!!!Nodule:", nodule_id, " has no charecteristics")
                continue
            if nodule.characteristics.malignancy is None:
                print("!!!!Nodule:", nodule_id, " has no malignacy")
                continue
         
            malignacy = nodule.characteristics.malignancy.text
            sphericiy = nodule.characteristics.sphericity.text
            margin = nodule.characteristics.margin.text
            spiculation = nodule.characteristics.spiculation.text
            texture = nodule.characteristics.texture.text
            calcification = nodule.characteristics.calcification.text
            internal_structure = nodule.characteristics.internalStructure.text
            lobulation = nodule.characteristics.lobulation.text
            subtlety = nodule.characteristics.subtlety.text

            #line = [nodule_id, x_center_perc, y_center_perc, z_center_perc, diameter_perc, malignacy]  
            line = [nodule_id, x_center_perc, y_center_perc, z_center_perc, diameter_perc, malignacy, sphericiy, margin, spiculation, texture, calcification, internal_structure, lobulation, subtlety]  
            extended_line = [ patient_id, nodule_id, x_center_perc, y_center_perc, z_center_perc, diameter_perc, malignacy, sphericiy, margin, spiculation, texture, calcification, internal_structure, lobulation, subtlety ]
                                                                                                   
            pos_lines.append(line)
            extended_lines.append(extended_line)
    

        nonNodules = reading_session.find_all("nonNodule")
        
        for nonNodule in nonNodules:
            
            z_center = float(nonNodule.imageZposition.text)
            z_center -= origin[2]
            z_center /= spacing[2]
            x_center = int(nonNodule.locus.xCoord.text)
            y_center = int(nonNodule.locus.yCoord.text)
            nodule_id = nonNodule.nonNoduleID.text
            x_center_perc = round(x_center / img_array.shape[2], 4)
            y_center_perc = round(y_center / img_array.shape[1], 4)
            z_center_perc = round(z_center / img_array.shape[0], 4)
            diameter_perc = round(max(6 / img_array.shape[2], 6 / img_array.shape[1]), 4)
    
            line = [nodule_id, x_center_perc, y_center_perc, z_center_perc, diameter_perc, 0]
            neg_lines.append(line)
    
    
    if agreement_threshold > 1:
        
        filtered_lines = []
        for pos_line1 in pos_lines:
            id1 = pos_line1[0]
            x1 = pos_line1[1]
            y1 = pos_line1[2]
            z1 = pos_line1[3]
            d1 = pos_line1[4]
            overlaps = 0
            for pos_line2 in pos_lines:
                id2 = pos_line2[0]
                if id1 == id2:
                    continue
                x2 = pos_line2[1]
                y2 = pos_line2[2]
                z2 = pos_line2[3]
                d2 = pos_line1[4]
                dist = math.sqrt(math.pow(x1 - x2, 2) + math.pow(y1 - y2, 2) + math.pow(z1 - z2, 2))
                if dist < d1 or dist < d2:
                    overlaps += 1
                    
            if overlaps >= agreement_threshold:
                filtered_lines.append(pos_line1)

        pos_lines = filtered_lines
        
    df_annos = pandas.DataFrame(pos_lines, columns=["anno_index", "coord_x", "coord_y", "coord_z", "diameter", "malscore" ,  "sphericiy", "margin", "spiculation", "texture", "calcification", "internal_structure", "lobulation", "subtlety"])
    df_annos.to_csv(LIDC_Annotations + "_labels/" + patient_id + "_annos_pos_lidc.csv", index=False)
    df_neg_annos = pandas.DataFrame(neg_lines, columns=["anno_index", "coord_x", "coord_y", "coord_z", "diameter", "malscore"])
    df_neg_annos.to_csv(LIDC_Annotations + "_labels/" + patient_id + "_annos_neg_lidc.csv", index=False)
      
    return pos_lines, neg_lines, extended_lines

def find_mhd_file(patient_id):
    for subject_no in range(settings.LUNA_SUBSET_START_INDEX, 10):
        src_dir = LUNA16_RAW_SRC_DIR  + "subset" + str(subject_no) + "/"
        for src_path in glob.glob(src_dir + "*.mhd"):
            if patient_id in src_path:
                return src_path
    return None
