==> step1_preprocess_make_train_cubes.ipynb <==
import settings
import helpers
import glob
import pandas
import ntpath
import numpy
import cv2
import os

CUBE_IMGTYPE_SRC = "_i"

if __name__ == "__main__":
    
    if not os.path.exists(settings.BASE_DIR_SSD + "generated_traindata/"):
        os.mkdir(settings.BASE_DIR_SSD + "generated_traindata/")

    if True:
        make_annotation_images_lidc()
    
    if False:
        make_annotation_images_manual()
  
    if False:
        make_candidate_auto_images(["falsepos", "luna", "edge"])
        
    if False:
        make_annotation_images_manual_ndsb3()  # for second model

# 主函数4

def make_annotation_images_manual_ndsb3():
    
    src_dir = "resources/ndsb3_manual_labels/"
    dst_dir = settings.BASE_DIR_SSD + "generated_traindata/ndsb3_train_cubes_manual/"
    
    if not os.path.exists(dst_dir):
        os.mkdir(dst_dir)
    
    #读取病人的label
    train_label_df = pandas.read_csv("resources/stage1_labels.csv")
    train_label_df.set_index(["id"], inplace=True)
    
    for file_path in glob.glob(dst_dir + "*.*"):
        os.remove(file_path)
    
    count=1
    
    #逐一遍历所有manual label的CSV文件
    for patient_index, csv_file in enumerate(glob.glob(src_dir + "*.csv")):
        
        print("Counter:",count)
        count+=1
        
        patient_id = ntpath.basename(csv_file).replace(".csv", "")
        df_annos = pandas.read_csv(csv_file)
        
        #如果没有label信息，则跳过
        if len(df_annos) == 0:
            continue
            
        #排除luna16数据集的数据
        if "1.3.6.1.4.1" in patient_id:
            continue
        
        #获取该病人的label
        cancer_label = train_label_df.loc[patient_id]["cancer"]
        
        #如果不存在该用户的数据，则跳过
        if not os.path.exists(settings.NDSB3_EXTRACTED_IMAGE_DIR + patient_id + "/"):
            continue
        
        #读取该病人重采样后的图片
        images = helpers.load_patient_images(patient_id, settings.NDSB3_EXTRACTED_IMAGE_DIR, "*" + CUBE_IMGTYPE_SRC + ".png")
        
        #计数器
        anno_index = 0
        
        #逐一遍历该病人的nodule label
        for index, row in df_annos.iterrows():
            
            #读取nodule信息
            pos_neg = "pos" if row["id"] == 0 else "neg"
            coord_x = int(row["x"] * images.shape[2])
            coord_y = int(row["y"] * images.shape[1])
            coord_z = int(row["z"] * images.shape[0])
            malscore = int(round(row["dmm"]))
            
            anno_index += 1
            
            #以nodule的点为中心，创建一个64*64*64的cube
            cube_img = get_cube_from_img(images, coord_x, coord_y, coord_z, 64)
            
            if cube_img.sum() < 5:
                print(" ***** Skipping ", coord_x, coord_y, coord_z)
                continue

            if cube_img.mean() < 10:
                print(" ***** Suspicious ", coord_x, coord_y, coord_z)

            if cube_img.shape != (64, 64, 64):
                print(" ***** incorrect shape !!! ", str(anno_index), " - ",(coord_x, coord_y, coord_z))
                continue
                
            print(patient_id)
            assert malscore > 0 or pos_neg == "neg"
            
            #将该cube数据以8*8个，64*64的图片形式保存
            save_cube_img(dst_dir + "ndsb3manual_" + patient_id + "_" + str(anno_index) + "_" + pos_neg + "_" + str(cancer_label) + "_" + str(malscore) + "_1_pn.png", cube_img, 8, 8)
        
        helpers.print_tabbed([patient_index, patient_id, len(df_annos)], [5, 64, 8])


# 主函数3

生成"falsepos", "luna", "edge"，三种false cube

def make_candidate_auto_images(candidate_types=[]):
    
    dst_dir = settings.BASE_DIR_SSD + "generated_traindata/luna16_train_cubes_auto/"
    
    if not os.path.exists(dst_dir):
        os.mkdir(dst_dir)

    for candidate_type in candidate_types:
        for file_path in glob.glob(dst_dir + "*_" + candidate_type + ".png"):
            os.remove(file_path)
    
    #对："falsepos", "luna", "edge"，分别进行处理        
    for candidate_type in candidate_types:
        
        if candidate_type == "falsepos":
            src_dir = "resources/luna16_falsepos_labels/"
        else:
            src_dir = settings.LUNA16_EXTRACTED_IMAGE_DIR + "_labels/"
        
        
        count=1
        
        #逐一遍历该类型的所有CSV文件
        for index, csv_file in enumerate(glob.glob(src_dir + "*_candidates_" + candidate_type + ".csv")):
            
            print("Candidate_Type:",candidate_type,"Counter:",count)
            count+=1
            
            patient_id = ntpath.basename(csv_file).replace("_candidates_" + candidate_type + ".csv", "")
            df_annos = pandas.read_csv(csv_file)
            #print(index, ",patient: ", patient_id, " type:", candidate_type) 
            
            #如果没有label信息，则跳过
            if len(df_annos) == 0:
                continue
            
            #读取该病人重采样后的图片
            images = helpers.load_patient_images(patient_id, settings.LUNA16_EXTRACTED_IMAGE_DIR, "*" + CUBE_IMGTYPE_SRC + ".png", exclude_wildcards=[])

            #计数器，记录该病人nodule数量
            row_no = 0
            
            #逐一遍历该病人的nodule label
            for index, row in df_annos.iterrows():
                
                #读取nodule信息
                coord_x = int(row["coord_x"] * images.shape[2])
                coord_y = int(row["coord_y"] * images.shape[1])
                coord_z = int(row["coord_z"] * images.shape[0])
                anno_index = int(row["anno_index"])
                
                #以nodule的点为中心，创建一个48*48*48的cube
                cube_img = get_cube_from_img(images, coord_x, coord_y, coord_z, 48)
                
                if cube_img.sum() < 10:
                    print("Skipping ", coord_x, coord_y, coord_z)
                    continue
               
                #将该cube数据以6*8个，48*48的图片形式保存
                try:       
                    save_cube_img(dst_dir + patient_id + "_" + str(anno_index) + "_0_" + candidate_type + ".png", cube_img, 6, 8)             
                except Exception as ex:
                    print(ex)

                #控制生成cube的数量上限
                row_no += 1
                max_item = 240 if candidate_type == "white" else 200
                if candidate_type == "luna":
                    max_item = 500
                if row_no > max_item:
                    break
                    

# 主函数2

根据nodule信息，获取64*64*64的cube

def make_annotation_images_manual():
    
    src_dir = "resources/luna16_manual_labels/"

    dst_dir = settings.BASE_DIR_SSD + "generated_traindata/luna16_train_cubes_manual/"
    
    if not os.path.exists(dst_dir):
        os.mkdir(dst_dir)

    for file_path in glob.glob(dst_dir + "*_manual.*"):
        os.remove(file_path)
    
    count=1
    
    #逐一遍历所有manual positive label文件
    for patient_index, csv_file in enumerate(glob.glob(src_dir + "*.csv")):
        
        patient_id = ntpath.basename(csv_file).replace(".csv", "") 
        df_annos = pandas.read_csv(csv_file)
        
        #检索某个具体病人的方法
        if "1.3.6.1.4" not in patient_id:
            continue
        
        #如果没有label信息，则跳过
        if len(df_annos) == 0:
            continue
       
        #读取该病人重采样后的图片
        images = helpers.load_patient_images(patient_id, settings.LUNA16_EXTRACTED_IMAGE_DIR, "*" + CUBE_IMGTYPE_SRC + ".png")
        
        #逐一遍历该病人的manual label
        for index, row in df_annos.iterrows():
            
            #读取nodule信息
            coord_x = int(row["x"] * images.shape[2])
            coord_y = int(row["y"] * images.shape[1])
            coord_z = int(row["z"] * images.shape[0])
            diameter = int(row["d"] * images.shape[2])
            node_type = int(row["id"])
            malscore = int(diameter)
            #malscore = min(25, malscore)
            #malscore = max(16, malscore)
            anno_index = index
            
            #以nodule的点为中心，创建一个64*64*64的cube
            cube_img = get_cube_from_img(images, coord_x, coord_y, coord_z, 64)
            
            if cube_img.sum() < 5:
                print(" ***** Skipping ", coord_x, coord_y, coord_z)
                continue

            if cube_img.mean() < 10:
                print(" ***** Suspicious ", coord_x, coord_y, coord_z)

            if cube_img.shape != (64, 64, 64):
                print(" ***** incorrect shape !!! ", str(anno_index), " - ",(coord_x, coord_y, coord_z))
                continue
            
            #将该cube数据以8*8个，64*64的图片形式保存
            save_cube_img(dst_dir + patient_id + "_" + str(anno_index) + "_" + str(malscore) + "_1_" + ("pos" if node_type == 0 else "neg") + ".png", cube_img, 8, 8)
        
        #输入各病人的cube数量
        helpers.print_tabbed([patient_index, patient_id, len(df_annos)], [5, 64, 8])
        
        print("Counter:",count)
        count+=1

# 主函数1

根据nodule信息，获取64*64*64的cube

def make_annotation_images_lidc():
    
    src_dir = settings.LUNA16_EXTRACTED_IMAGE_DIR + "_labels/"
    
    dst_dir = settings.BASE_DIR_SSD + "generated_traindata/luna16_train_cubes_lidc/"
    
    if not os.path.exists(dst_dir):
        os.mkdir(dst_dir)

    for file_path in glob.glob(dst_dir + "*.*"):
        os.remove(file_path)
    
    count=1
    
    #逐一遍历所有病人的positive label文件
    for patient_index, csv_file in enumerate(glob.glob(src_dir + "*_annos_pos_lidc.csv")):
        
        patient_id = ntpath.basename(csv_file).replace("_annos_pos_lidc.csv", "") 
        df_annos = pandas.read_csv(csv_file)
        
        #如果没有label信息，则跳过
        if len(df_annos) == 0:
            continue
        
        #读取该病人重采样后的图片
        images = helpers.load_patient_images(patient_id, settings.LUNA16_EXTRACTED_IMAGE_DIR, "*" + CUBE_IMGTYPE_SRC + ".png")
        
        #逐一遍历该病人的positive label
        for index, row in df_annos.iterrows():
            
            #读取nodule信息
            coord_x = int(row["coord_x"] * images.shape[2])
            coord_y = int(row["coord_y"] * images.shape[1])
            coord_z = int(row["coord_z"] * images.shape[0])
            malscore = int(row["malscore"])
            anno_index = row["anno_index"]
            anno_index = str(anno_index).replace(" ", "xspacex").replace(".", "xpointx").replace("_", "xunderscorex")
       
            #以nodule的点为中心，创建一个64*64*64的cube
            cube_img = get_cube_from_img(images, coord_x, coord_y, coord_z, 64)
        
            if cube_img.sum() < 5:
                print(" ***** Skipping ", coord_x, coord_y, coord_z)
                continue

            if cube_img.mean() < 10:
                print(" ***** Suspicious ", coord_x, coord_y, coord_z)

            if cube_img.shape != (64, 64, 64):
                print(" ***** incorrect shape !!! ", str(anno_index), " - ",(coord_x, coord_y, coord_z))
                continue
            
            #将该cube数据以8*8个，64*64的图片形式保存
            save_cube_img(dst_dir + patient_id + "_" + str(anno_index) + "_" + str(malscore) + "_1_pos.png", cube_img, 8, 8)
        
        #输入各病人的cube数量
        helpers.print_tabbed([patient_index, patient_id, len(df_annos)], [5, 64, 8])
           
        print("Counter:",count)
        count+=1

def get_cube_from_img(img3d, center_x, center_y, center_z, block_size):
    start_x = max(center_x - block_size / 2, 0)
    if start_x + block_size > img3d.shape[2]:
        start_x = img3d.shape[2] - block_size

    start_y = max(center_y - block_size / 2, 0)
    start_z = max(center_z - block_size / 2, 0)
    if start_z + block_size > img3d.shape[0]:
        start_z = img3d.shape[0] - block_size
    start_z = int(start_z)
    start_y = int(start_y)
    start_x = int(start_x)
    res = img3d[start_z:start_z + block_size, start_y:start_y + block_size, start_x:start_x + block_size]
    return res

def save_cube_img(target_path, cube_img, rows, cols):
    assert rows * cols == cube_img.shape[0]
    img_height = cube_img.shape[1]
    img_width = cube_img.shape[1]
    res_img = numpy.zeros((rows * img_height, cols * img_width), dtype=numpy.uint8)

    for row in range(rows):
        for col in range(cols):
            target_y = row * img_height
            target_x = col * img_width
            res_img[target_y:target_y + img_height, target_x:target_x + img_width] = cube_img[row * cols + col]

    cv2.imwrite(target_path, res_img)

int(4.6)
