==> Step3_Recall_Rate_Calculation.ipynb <==
import pandas
import settings
import helpers
import random
import os
import glob
import ntpath
import math

CUBE_IMGTYPE_SRC = "_i"

#O 代表scanner的输出，T代表真实值

# Output from Scanner
O_dir = "D:/jupyter-notebook/LungCancerPredict/NMS_Result/CNN_10_9/"
#D:\jupyter-notebook\LungCancerPredict\NMS_Result\RNN_10_9
#D:\jupyter-notebook\LungCancerPredict\NMS_Result\CNN_10 （NMS的召回率）
#D:\jupyter-notebook\LungCancerPredict\LIDC_Annotations\_labels   （lidc的在luna里的包含率）
#D:\jupyter-notebook\LungCancerPredict\extracted\luna16_nodule_predictions_RNN\predictions10_luna16_fs (scnner的召回率)

# True from Radiologists
#T_dir =  "D:/jupyter-notebook/LungCancerPredict/LUNA_Annotations/pos_from_luna/"
T_dir = "D:/jupyter-notebook/LungCancerPredict/LUNA_Annotations/pos_from_luna/"

#统计没有positive的患者数量
count1=0
#统计recall的总量
count2=0
#统计true postive的总量
count3=0
#统计Output的总量
count4=0
#统计一共读取了几个postive文件
count5=0

for patient_index, csv_file in enumerate(glob.glob(O_dir + "*.csv")):
    
    patient_id = ntpath.basename(csv_file).replace("_annos_pos_lidc.csv", "") 
    #patient_id = ntpath.basename(csv_file).replace(".csv", "") 
    #patient_img =helpers.load_patient_images(patient_id, settings.LUNA16_EXTRACTED_IMAGE_DIR, "*_i.png", [])
    #查看病人ID
    #print(patient_id)
    
    #读取该病人的Output数据
    df_O = pandas.read_csv(csv_file)

    #统计Output的总量
    count4+=len(df_O)
    
    #寻找与当前病人匹配的15文件
    O_dir_15 = settings.BASE_DIR_SSD + "extracted/luna16_nodule_predictions_CNN/predictions15_luna16_fs/"
    df_O_15=[]
    for patient_index, csv_file in enumerate(glob.glob(O_dir_15 + "*.csv")):
        
        patient_id_15 = ntpath.basename(csv_file).replace(".csv", "") 
        
        if patient_id_15 == patient_id:
            
            df_O_15=pandas.read_csv(csv_file)
            #count4+=len(df_O_15)

            break
    
    #寻找与当前病人匹配的20文件
    O_dir_20 = settings.BASE_DIR_SSD + "extracted/luna16_nodule_predictions_CNN/predictions20_luna16_fs/"
    df_O_20=[]
    for patient_index, csv_file in enumerate(glob.glob(O_dir_20 + "*.csv")):
        
        patient_id_20 = ntpath.basename(csv_file).replace(".csv", "") 
        
        if patient_id_20 == patient_id:
            
            df_O_20=pandas.read_csv(csv_file)
            #count4+=len(df_O_20)
            break
            
    #df_O.append(df_O_15)
    #df_O.append(df_O_20)

    
    #寻找与当前病人匹配的True Positive文件
    T_dir =  T_dir
    df_T=[]
    for patient_index, csv_file in enumerate(glob.glob(T_dir + "*_annos_pos_lidc.csv")):
        
        patient_id_T = ntpath.basename(csv_file).replace("_annos_pos_lidc.csv", "") 
        
        if patient_id_T == patient_id:
            
            df_T=pandas.read_csv(csv_file)
            count3+=len(df_T)
            count5+=1
            break
        
    #如果该病人没有positive则跳过
    if df_T.empty==True:
        count1+=1
        continue
        
    if df_O.empty==True:
        continue

    #计算与True的距离
    distance=[]
    
    for index_T, row_T in df_T.iterrows():
        
        for index_O, row_O in df_O.iterrows():
            
            d = calculateD(row_O,row_T)
            distance.append(d)
        
        #threshold = row_T["diameter"]
        #threshold = math.sqrt((16**2+16**2+16**2))/patient_img.shape[2]
        threshold=0.05
        
        if min(distance)<threshold:
            count2+=1
        print(count2)

print(count2)
print(count3)
print(count4)
print(count5)

def calculateD(data1,data2):
     
    d = (data1["coord_x"]-data2["coord_x"])**2 + (data1["coord_y"]-data2["coord_y"])**2 +(data1["coord_z"]-data2["coord_z"])**2  
    d = math.sqrt(d)
    
    return d
