==> Step4_fix_probability.ipynb <==
import pandas
import numpy as np

df = pandas.read_csv("D:/jupyter-notebook/LNDb/result_c/temp.csv")

rater_a = df.iloc[:,5].values.tolist()
for i in range(len(df)):
    rater_a[i] = calc_texture_class(rater_a[i])

rater_b = []
rater_b_temp = df.iloc[:,2:5].values.tolist()
for i in range(len(df)):
    max_index = rater_b_temp[i].index(max(rater_b_temp[i]))
    rater_b.append(max_index)

rater_b_temp = np.array(rater_b_temp)
print(rater_b_temp[0])
rater_b_temp = rater_b_temp*[1.5,1.5,0.6]
print(rater_b_temp[0])
rater_b_temp = rater_b_temp.tolist()

df

def calc_texture_class(texture):
    
    threshold = [7/3,11/3]
   
    if texture>=threshold[0] and texture<=threshold[1]:
        vclass = 1
    elif texture>threshold[1]:
        vclass = 2
    else:
        vclass = 0
 
    return vclass
