==> Step4_Process_Predicted_Annotations.ipynb <==
import settings
import helpers
import SimpleITK  # conda install -c https://conda.anaconda.org/simpleitk SimpleITK
import numpy
import pandas
import ntpath

import cv2  # conda install -c https://conda.anaconda.org/menpo opencv3
import shutil
import random
import math
import multiprocessing
from bs4 import BeautifulSoup #  conda install beautifulsoup4, coda install lxml
import os
import glob

if __name__ == "__main__":
    process_predited_annotations()

def process_predited_annotations():
                      
    
    ndsb_list = os.listdir(settings.NDSB3_EXTRACTED_IMAGE_DIR)
    anno_list = pandas.read_csv("D:/jupyter-notebook/LungCancerPredict/NDSB_Annotations_From_CNN/ResNet/out.csv")
    
    count=0
    
    for patient_id in ndsb_list:
        
        df = anno_list.loc[anno_list["patient_id"]==patient_id ]
        
        df.to_csv("D:/jupyter-notebook/LungCancerPredict/NDSB_Annotations_From_CNN/ResNet/_labels_predicted/" + patient_id + ".csv", index=False)
        
        count+=1
        print(count)
